Move `app/service/session` to `addon`
According to [Dockyard's Blog Post][dockyard] (section #2):

> A consumer should have the ability to easily extend your addon
> to do whatever they want. This means organizing your code a
> certain way.
> To provide this you should put all of your business logic into
> `addon/` and then include wrapper classes in `app/` that just `import`
> then `export` the extended class.

[dockyard]: https://dockyard.com/blog/2015/03/22/tips-for-writing-ember-addons
